refactor similar to geo setup
