--- src/kindtech/ons.py
@@ -1,20 +1,0 @@
-"""Office for National Statistics (ONS) data loading and processing module."""
-
-
-def load_ons(dataset=None, period=None, geography=None):
-    """
-    Load data from the Office for National Statistics.
-
-    Args:
-        dataset (str, optional): The ONS dataset identifier.
-        period (str, optional): Time period for the data (e.g., '2023', 'Q1-2023').
-        geography (str, optional): Geographic area code or name.
-
-    Returns:
-        dict: Loaded ONS data.
-    """
-    # Implementation placeholder
-    if not dataset:
-        raise ValueError("Dataset identifier must be provided")
-
-    return {"dataset": dataset, "period": period, "geography": geography, "data": {}}
