+++ src/kindtech/ons/__init__.py
@@ -1,0 +1,5 @@
+"""ONS data loading and processing package."""
+
+from .api import load_ons
+
+__all__ = ["load_ons"]
