+++ src/kindtech/ons/_ingestion.py
@@ -1,13 +1,34 @@
-"""NOMIS API client for accessing UK official statistics."""
+"""
+Nomis tables ingestion module
 
-from io import StringIO
+This module provides functions to load NOMIS tables from the nomis API and save
+it to a CSV file
+
+The CSV file is later used as the main lookup table to find the relevant data
+
+More information on the NOMIS API can be found here:
+https://www.nomisweb.co.uk/api/v01/home.html
+"""
+
+import logging
+from pathlib import Path
 from typing import Any
 
 import pandas as pd
 import requests
 
+# Set up logging
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+)
+logger = logging.getLogger(__name__)
 
-def list_tables(base_url: str = "https://www.nomisweb.co.uk/api/v01") -> pd.DataFrame:
+# Constants
+NOMIS_BASE_URL = "https://www.nomisweb.co.uk/api/v01"
+DEFAULT_OUTPUT_PATH = Path(__file__).parent / "data" / "nomis_tables.csv"
+
+
+def list_tables(base_url: str = NOMIS_BASE_URL) -> pd.DataFrame:
     """
     Fetch available NOMIS datasets and return them as a DataFrame.
 
@@ -43,7 +64,7 @@ def list_tables(base_url: str = "https://www.nomisweb.co.uk/api/v01") -> pd.Data
 
 
 def list_data_sources(
-    base_url: str = "https://www.nomisweb.co.uk/api/v01",
+    base_url: str = NOMIS_BASE_URL,
 ) -> pd.DataFrame:
     """
     List NOMIS data sources.
@@ -98,105 +119,7 @@ def list_data_sources(
     return pd.DataFrame(sources_list)
 
 
-def get_ons_table(
-    dataset_id: str, base_url: str = "https://www.nomisweb.co.uk/api/v01", **kwargs
-) -> pd.DataFrame:
-    """
-    Retrieve data from the Office of National Statistics (ONS) via the NOMIS API
-    for a given dataset.
-
-    Args:
-        dataset_id: A table ID recognised by NOMIS (e.g. "NM_1_1")
-        base_url: NOMIS base URL to query
-        **kwargs: Additional parameters used to filter and aggregate the data
-                 retrieved from ONS. These are passed to build_url_query_string
-                 to build everything in the URL after "?". See
-                 "https://www.nomisweb.co.uk/api/v01/help" for information on
-                 the options available
-
-    Returns:
-        A DataFrame containing the data downloaded from ONS' NOMIS API
-
-    Examples:
-        >>> # Pull all data from the "Jobseeker's Allowance" dataset
-        >>> get_ons_table("NM_1_1")
-
-        >>> # Filter "Jobseeker's Allowance" dataset and select columns to output
-        >>> get_ons_table("NM_1_1", geography="TYPE480", time="latest",
-        ...               measures=20100, item=1,
-        ...               select=["geography_name", "sex_name", "obs_value"])
-
-        >>> # Aggregate statistics using 'rows' and 'cols'
-        >>> get_ons_table("NM_1_1", geography="TYPE480", time="latest",
-        ...               measures=20100, item=1,
-        ...               select=["geography_name", "sex_name", "obs_value"],
-        ...               rows=["geography_name"], cols=["sex_name"])
-    """
-    # Build query string from kwargs
-    query_string = build_url_query_string(**kwargs)
-
-    # Construct the full URL
-    nomis_url = f"{base_url}/dataset/{dataset_id}.data.csv{query_string}"
-
-    print(f"Querying NOMIS API -> {nomis_url}")
-
-    # Make the GET request
-    response = requests.get(nomis_url)
-    response.raise_for_status()
-
-    # Check if response is HTML (error page)
-    if response.text.strip().startswith("<!DOCTYPE html>"):
-        raise ValueError(
-            "NOMIS ID does not exist. Use list_tables() for full list of IDs"
-        )
-
-    # Parse CSV data
-    try:
-        nomis_table = pd.read_csv(StringIO(response.text))
-    except Exception as e:
-        raise ValueError(f"Failed to parse CSV data: {e}") from e
-
-    # Check for truncation warning
-    if len(nomis_table) == 25000:
-        print(
-            "Warning: Query has been truncated to 25,000 rows - providing a NOMIS UID "
-            "(i.e. uid = '0x...') will return the full table. See 'API Key, your "
-            "Unique ID and Signatures' section at "
-            "https://www.nomisweb.co.uk/api/v01/help for further details"
-        )
-
-    return nomis_table
-
-
-def build_url_query_string(**kwargs) -> str:
-    """
-    Build URL query string from keyword arguments.
-
-    Args:
-        **kwargs: Parameters to include in the query string
-
-    Returns:
-        Query string starting with "?" if parameters exist, empty string otherwise
-    """
-    if not kwargs:
-        return ""
-
-    # Convert parameters to query string format
-    query_parts = []
-    for key, value in kwargs.items():
-        if isinstance(value, list | tuple):
-            # Handle list/tuple parameters (like select, rows, cols)
-            for item in value:
-                query_parts.append(f"{key}={item}")
-        else:
-            query_parts.append(f"{key}={value}")
-
-    return "?" + "&".join(query_parts)
-
-
-def get_overview(
-    dataset_id: str, base_url: str = "https://www.nomisweb.co.uk/api/v01"
-) -> dict[str, Any]:
+def get_overview(dataset_id: str, base_url: str = NOMIS_BASE_URL) -> dict[str, Any]:
     """
     Get overview information for a specific NOMIS dataset.
 
@@ -236,7 +159,7 @@ def is_nested(lst: list) -> bool:
 
 
 def extract_data_source(
-    base_url: str = "https://www.nomisweb.co.uk/api/v01",
+    base_url: str = NOMIS_BASE_URL,
 ) -> pd.DataFrame:
     """
     Extract data sources for all NOMIS tables (excluding specific ones).
@@ -317,7 +240,7 @@ def extract_data_source(
 
 
 def create_nomis_tables_dataset(
-    base_url: str = "https://www.nomisweb.co.uk/api/v01",
+    base_url: str = NOMIS_BASE_URL,
 ) -> pd.DataFrame:
     """
     Create a comprehensive NOMIS tables dataset with sources.
@@ -342,7 +265,7 @@ def create_nomis_tables_dataset(
 
 def save_nomis_tables_dataset(
     filepath: str = "nomis_tables.csv",
-    base_url: str = "https://www.nomisweb.co.uk/api/v01",
+    base_url: str = NOMIS_BASE_URL,
 ) -> None:
     """
     Create and save the NOMIS tables dataset to a file.
